requirements.txt
flask>=2.3.0
pillow>=10.0.0
requests>=2.31.0
pygame>=2.5.0
python-dotenv>=1.0.0
psutil>=5.9.0

